pyproject.toml
[tool.poetry]
name = "onpoint-kb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.9.7 || >3.9.7,<4.0"
nltk = "^3.8.1"
pinecone-client = "^2.2.4"
langchain = "^0.0.320"
google-cloud-aiplatform = ">=1.35.0"
streamlit = "^1.27.2"
python-dotenv = "^1.0.0"
streamlit-chat = "^0.1.1"
python-slugify = "^8.0.1"
streamlit-option-menu = "^0.3.6"
pydantic-settings = "^2.0.3"
pydantic = "^2.4.2"
google-cloud-bigquery = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
